PlayerFactoryService

+++ Top-Down/Assets/GameAssets/Scripts/Enemy/EnemyLevelInfo.cs
@@ -8,7 +8,8 @@ namespace TopDown {
     [field: SerializeField] public float AttackRadius { get; private set; }
 
     [field: SerializeField] public float MoveSpeed { get; private set; }
-    [field: SerializeField] public float Health { get; private set; }
-    [field: SerializeField] public float Damage { get; private set; }
+    [field: SerializeField] public int Health { get; private set; }
+    
+    [field: SerializeField] public int Damage { get; private set; }
   }
 }
